requirements.txt
sagemaker-inference
numpy
pandas
scikit-learn
joblib
boto3
snowflake-connector-python[pandas]
pyarrow